Pause Menu Prefab
Okay the Pause Menu Prefab needs to be added to the HUD 
The Pause Menu Controller just needs to be added to the hierarchy with the following things added to the inspector:
   Pause Menu UI = PauseMenuMain
   Pause Screen = Pause Background
   How To Play = How To Play

The Pause Menu Controller needs to be put into the None (Object) spot in the bottom of the Inspector for the Resume Button, then (No Function) will be changed to PauseMenu.DeactivateMenu from the Pause Menu drop down. 

The Pause Menu Controller needs to be put into the None (Object) spot in the bottom of the Inspector for the How to Play Button, then (No Function) will be changed to PauseMenu.HowTo() from the Pause Menu drop down. 

The Pause Menu Controller needs to be put into the None (Object) spot in the bottom of the Inspector for the Return Button, then (No Function) will be changed to PauseMenu.Return from the Pause Menu drop down.